Remove CNRS stuff Add some aliases (clean & autoremove) Add exception for docker (WIP)

+++ .zshrc
@@ -4,31 +4,8 @@
 source ~/.zshrc.oh-my-zsh
 READNULLCMD=${PAGER:-/usr/bin/pager}
 
-# Load Grid environment
-if [[ -f /etc/profile.d/grid-env.sh ]]; then
-	source /etc/profile.d/grid-env.sh
-	export PATH=/opt/glite/yaim/bin:$PATH
-	alias gjc='glite-wms-job-submit -a'
-	alias gjs='glite-wms-job-status'
-fi
-
-export X509_CERT_DIR=/etc/grid-security/certificates
-export X509_USER_CERT=~/.globus/usercert.pem
-export X509_USER_KEY=~/.globus/userkey.pem
-export X509_USER_PROXY=/tmp/x509up_u500
-
-# Load openstask testing creds
-if [[ -f ~/scripts/creds ]]; then
-        source ~/scripts/creds
-fi
-
-# SCM @ IN2P3
-export SVN_SSH=ssh
-
-# Lauch screen
-if [[ `hostname` == "vinc" ]]; then
-	screen -r irc
-fi
+# Add Puppet in PATH
+export PATH=$PATH:/opt/puppetlabs/bin
 
 # Aliases
 alias h=' search_history'
@@ -53,7 +30,8 @@ alias os="openstack"
 alias osi="openstack --insecure"
 
 # debian
-alias acs='apt-cache search'
+#alias acs='apt-cache search'
+alias acs='aptitude search'
 alias acss='apt-cache show'
 alias ags='apt-get source'
 alias agU='sudo apt-get update'
@@ -63,8 +41,9 @@ alias agi='sudo apt-get install'
 alias agrm='sudo apt-get remove'
 alias agrmp='sudo apt-get remove --purge'
 alias ag='sudo apt-get'
-alias acs='aptitude search'
 alias acp='apt-cache policy'
+alias agar='sudo apt-get autoremove'
+alias agc='sudo apt-get clean'
 
 # redhat/centos
 alias yu='sudo yum update'
@@ -76,9 +55,11 @@ alias ss='sudo systemctl status'
 alias sdr='sudo systemctl daemon-reload'
 
 # docker
-alias do='docker'
-alias dops='docker ps -a'
-alias dorm='docker rm'
+if [[ -e /usr/bin/docker ]]; then
+   alias do='docker'
+   alias dops='docker ps -a'
+   alias dorm='docker rm'
+fi
 
 # Global ZSH
 [ -x "/usr/bin/most" ] && export PAGER=most
